+++ mithril-stm/src/aggregate_signature/signature.rs
@@ -131,19 +131,24 @@ impl<D: Clone + Digest + FixedOutput + Send + Sync> StmAggrSig<D> {
 
     /// Convert multi signature to bytes
     /// # Layout
+    /// * Aggregate signature type (u8)
     /// * Number of the pairs of Signatures and Registered Parties (SigRegParty) (as u64)
-    /// * Size of a pair of Signature and Registered Party
-    /// * Pairs of Signatures and Registered Parties
+    /// * Pairs of Signatures and Registered Parties (prefixed with their size as u64)
     /// * Batch proof
     pub fn to_bytes(&self) -> Vec<u8> {
         let mut out = Vec::new();
+        // This proof type is not strictly necessary, but it will help to identify
+        // the type of the proof used to aggregate when implementing multiple aggregation proof systems.
+        // We use '0' for concatenation proof.
+        let proof_type: u8 = 0;
+        out.extend_from_slice(&proof_type.to_be_bytes());
         out.extend_from_slice(&u64::try_from(self.signatures.len()).unwrap().to_be_bytes());
-        out.extend_from_slice(
-            &u64::try_from(self.signatures[0].to_bytes().len())
-                .unwrap()
-                .to_be_bytes(),
-        );
         for sig_reg in &self.signatures {
+            out.extend_from_slice(
+                &u64::try_from(sig_reg.to_bytes().len())
+                    .unwrap()
+                    .to_be_bytes(),
+            );
             out.extend_from_slice(&sig_reg.to_bytes());
         }
         let proof = &self.batch_proof;
@@ -154,26 +159,53 @@ impl<D: Clone + Digest + FixedOutput + Send + Sync> StmAggrSig<D> {
 
     ///Extract a `StmAggrSig` from a byte slice.
     pub fn from_bytes(bytes: &[u8]) -> Result<StmAggrSig<D>, StmAggregateSignatureError<D>> {
+        let mut u8_bytes = [0u8; 1];
+        let mut bytes_index = 0;
+
+        u8_bytes.copy_from_slice(
+            bytes
+                .get(..1)
+                .ok_or(StmAggregateSignatureError::SerializationError)?,
+        );
+        bytes_index += 1;
+        let proof_type = u8::from_be_bytes(u8_bytes);
+        if proof_type != 0 {
+            return Err(StmAggregateSignatureError::SerializationError);
+        }
+
         let mut u64_bytes = [0u8; 8];
-
-        u64_bytes.copy_from_slice(&bytes[..8]);
-        let size = usize::try_from(u64::from_be_bytes(u64_bytes))
+        u64_bytes.copy_from_slice(
+            bytes
+                .get(bytes_index..bytes_index + 8)
+                .ok_or(StmAggregateSignatureError::SerializationError)?,
+        );
+        let total_sigs = usize::try_from(u64::from_be_bytes(u64_bytes))
             .map_err(|_| StmAggregateSignatureError::SerializationError)?;
+        bytes_index += 8;
 
-        u64_bytes.copy_from_slice(&bytes[8..16]);
-        let sig_reg_size = usize::try_from(u64::from_be_bytes(u64_bytes))
-            .map_err(|_| StmAggregateSignatureError::SerializationError)?;
-
-        let mut sig_reg_list = Vec::with_capacity(size);
-        for i in 0..size {
+        let mut sig_reg_list = Vec::with_capacity(total_sigs);
+        for _ in 0..total_sigs {
+            u64_bytes.copy_from_slice(
+                bytes
+                    .get(bytes_index..bytes_index + 8)
+                    .ok_or(StmAggregateSignatureError::SerializationError)?,
+            );
+            let sig_reg_size = usize::try_from(u64::from_be_bytes(u64_bytes))
+                .map_err(|_| StmAggregateSignatureError::SerializationError)?;
             let sig_reg = StmSigRegParty::from_bytes::<D>(
-                &bytes[16 + (sig_reg_size * i)..16 + (sig_reg_size * (i + 1))],
+                bytes
+                    .get(bytes_index + 8..bytes_index + 8 + sig_reg_size)
+                    .ok_or(StmAggregateSignatureError::SerializationError)?,
             )?;
+            bytes_index += 8 + sig_reg_size;
             sig_reg_list.push(sig_reg);
         }
 
-        let offset = 16 + sig_reg_size * size;
-        let batch_proof = BatchPath::from_bytes(&bytes[offset..])?;
+        let batch_proof = BatchPath::from_bytes(
+            bytes
+                .get(bytes_index..)
+                .ok_or(StmAggregateSignatureError::SerializationError)?,
+        )?;
 
         Ok(StmAggrSig {
             signatures: sig_reg_list,